requirements.txt
discord.py==2.2.3
aiosqlite==0.18.0
python-dotenv==1.0.0
requests==2.28.2 